Creates ListWrapper
